% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppar.R
\docType{data}
\name{bcm}
\alias{bcm}
\title{Breast cancer metastases from different anatomical sites}
\format{Contains gene expression matrix, phenotype (pData) and feature (fData) data:
\describe{
  \item{ID}{In fData(e) -- probe IDs}
  \item{EntrezGeneID}{In fData(e) -- Entrez Ids}
  \item{GeneSymbol}{In fData(e) -- Gene Symbols}
  \item{geo_accession}{In pData(e) -- GEO accession IDs}
  \item{source_name_ch1}{In pData(e) -- Sample information}
}}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE46141}
}
\usage{
data(GSE46141)
}
\value{
An ExpressionSet object
}
\description{
An ExpressionSet object containing trimmed GSE46141 data. The object contains
gene expression measurements on local breast tumours and liver, lymph node, skin local-
regional etc metastatic tumours. Contains 9503 features and 80 samples (ascite, bone, lung and skin
samples were removed).
}
\keyword{datasets}

