% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppar.R
\docType{data}
\name{maupin}
\alias{maupin}
\title{Maupin's TGFb data and a TGFb gene signature}
\format{A list of two components:
\describe{
  \item{M_Ctrl_R1}{In data; Control Sample - Replicate 1}
  \item{M_Ctrl_R2}{In data; Control Sample - Replicate 2}
  \item{M_Ctrl_R3}{In data; Control Sample - Replicate 3}
  \item{M_TGFb_R1}{In data; Control Sample - Replicate 1}
  \item{M_TGFb_R2}{In data; Control Sample - Replicate 2}
  \item{M_TGFb_R3}{In data; Control Sample - Replicate 3}
  \item{EntrezID}{In sig; The Entrez IDs}
  \item{Symbol}{In sig; Gene symbols}
  \item{upDown}{In sig; Direction of regulation e.g. up, down}

  }}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE23952}
}
\usage{
data(Maupin)
}
\value{
A list of 2
}
\description{
A list consisting of two components:
data: A matrix consisting of gene expression values on 3 control
and 3 TGFb treated samples.
sig: A TGFb gene signature. A dataframe containing a gene signature
and information on whether genes in the signature are up or dow regulated.
}
\keyword{datasets}

