% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppar.R
\docType{package}
\name{oppar}
\alias{oppar}
\alias{oppar-package}
\title{oppar: A package for outlier profile and pathway analysis.}
\description{
The oppar package provides 3 main function for outlier profile analysis:
opa, getSampleOutlier and getSubtypeProbes, and 1 function for pathway and
gene set enrichment analysis, based on gsva function implemented in the GSVA package
}
\section{opa}{

calculates the outlier profile matrix, using the method proposed in Wang et al. (2012) paper
}

\section{getSampleOutlier}{

extracts outlier profile for individual samples
}

\section{getSubtypeProbes}{

extracts outlier profile for a group of related samples, such as subtypes
}

