% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_OPPARList.R
\docType{class}
\name{OPPARList-class}
\alias{$,OPPARList-method}
\alias{OPPARList-class}
\alias{show,OPPARList-method}
\title{A S4 class for the output of OPPAR main function, opa.}
\usage{
\S4method{show}{OPPARList}(object)

\S4method{$}{OPPARList}(x, name)
}
\arguments{
\item{object}{An object of type OPPARList}

\item{x}{Object of type OPPARList.}

\item{name}{Name of the slot to access.}
}
\value{
returns the number of outlier features detected, the number of samples retained,
    and the parameters used to run the \code{opa} function

extracts slots from an object of type \code{OPPARList}.
}
\description{
A S4 class for the output of OPPAR main function, opa.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: A show method for objects of class OPPARList

\item \code{$}: A method to extract slots in \code{OPPARList}
}}
\section{Slots}{

\describe{
\item{\code{profileMatrix}}{A matrix of 0, -1 and 1 representing outlier genes in samples}

\item{\code{upper.quantile}}{Numeric. The upper quantile cut-off for detection of outliers}

\item{\code{lower.quantile}}{Numeric. The lower quantile cut-off for detection of outliers}

\item{\code{group}}{A factor vector representing the group to which each sample belong}

\item{\code{.Data}}{matrix}
}}

